+++ lib/src/api/bytelist.dart
@@ -75,7 +75,7 @@ class ByteList with ListMixin<int> implements Uint8List {
 
   @override
   ByteList sublist(int start, [int end]) {
-      final sublist = _u8l.sublist(start, end);
-      return ByteList(sublist, sublist.length);
+    final sublist = _u8l.sublist(start, end);
+    return ByteList(sublist, sublist.length);
   }
 }